a kind of not really but at least somewhat cleanroomed version of the fixes from #74 to be perfectly honest i mostly just went through and double checked that everything seemed okay, because once the code is written you can't exactly put it back in the box

+++ src/main/java/dev/latvian/mods/rhino/NativeString.java
@@ -856,7 +856,7 @@ final class NativeString extends IdScriptableObject implements Wrapper {
 					if (id == Id_charAt) {
 						return String.valueOf(c);
 					}
-					return c;
+					return (int) c; // js has no char type, so return as an int
 				}
 
 				case Id_indexOf: {
